Rewrite grammar to use internal node representation
This is the first (well second, considering that we removed the
dependency on xmlList for definitions in the previous commit) commit to
remove the dependency on libxml2.  This is both an experiment to see how
simple and efficient it could be made and also to remove all external
dependencies beyond libc.  I think it proved to be a net gain, even
though the anchor/footnote/auxiliary stuff is a bit convoluted.