Add private anchor_node instead of having anchor in node
This’ll be followed by further changes to clean up node and the public
interface.  I’m not even sure node should be part of said interface, but
it’s better to have this kind of polymorphism anyway.  The anchor_node
handling is a lot cleaner than what we previously had.