Improve eol rules to properly support nested lists
We do this by adding some flexibility to the rules.  We now allow tight
sublists such as

•   A
  •   B

and add support for starting a paragraph with a bol symbol if it’s
preceded by an explicit paragraph symbol (¶).  The handling of indents
and dedents is further made more flexible by assuming that the author
knows what they’re doing and treating lines that would otherwise be
errors as deindented paragraphs.