Fix buffer node creation
The previous code assumed that we were ever only creating one buffer at
a time.  While strictly true, we need to textify that buffer before we
continue on the next one.  This was done, but if we encounter a group, a
new buffer may be created, in which case we won’t textify until it’s too
late.  We now keep track of the buffer node so that we may textify it
whenever we start to build a new buffer.